generic await impl
